+++ test/integration/tests/features/cssproperties.mjs
@@ -16,7 +16,7 @@ const testSteps = [
     `
     document.head.appendChild(style)
 
-    chart.feature('css-properties', true)
+    chart.feature('cssProperties', true)
 
     //    const body = document.querySelector('body')
     //    body.style.setProperty('--vizzu-plot-yAxis-label-color', '#ff0000')